requirements.txt
streamlit==1.40.0
boto3==1.35.0
pymysql==1.1.1
python-dotenv==1.0.1
cryptography
